+++ package.json
@@ -28,8 +28,8 @@
         "react-redux": "^9.1.0",
         "react-router-dom": "^6.22.0",
         "react-scripts": "^5.0.1",
-        "react-super-responsive-table": "^5.2.3",
         "react-spring": "^9.7.3",
+        "react-super-responsive-table": "^5.2.3",
         "react-toastify": "^9.1.3",
         "react-type-animation": "^3.2.0",
         "recharts": "^2.12.6",
