+++ src/Components/Dashboard/MyProfile/MyPreferenceSection.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const MyPreferenceSection = () => {
+  return <div>MyPreferenceSection</div>;
+};
+
+export default MyPreferenceSection;
