--- src/Components/Profile/AboutSection.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const AboutSection = () => {
-  return <div></div>;
-};
-
-export default AboutSection;
